--- j2/platformio.ini
@@ -1,36 +1,0 @@
-; PlatformIO Project Configuration File
-;
-;   Build options: build flags, source filter
-;   Upload options: custom upload port, speed and extra flags
-;   Library options: dependencies, extra library storages
-;   Advanced options: extra scripting
-;
-; Please visit documentation for the other options and examples
-; https://docs.platformio.org/page/projectconf.html
-
-[env:esp32-s3-devkitc-1]
-platform = espressif32
-; platform = https://github.com/platformio/platform-espressif32.git
-board = esp32-s3-devkitc-1
-framework = arduino
-
-monitor_speed = 115200
-monitor_raw = true
-
-lib_deps = 
-    rover_can_lib=symlink://../lib/rover_can_lib
-    lib_rover=symlink://../lib/lib_rover
-
-build_unflags =
-    ; -w CCFLAGS
-    -std=gnu++11
-
-build_flags =
-    -Wall
-    -Wextra
-    
-    '-D ARDUINO_USB_CDC_ON_BOOT=1'
-    '-D ARDUINO_USB_MODE=1'
-
-    '-D VERBOSE'
-    '-D LOGGER_LOWEST_LEVEL=0'
